refactor: reorganize environment variable handling and include necessary headers
