+++ src/builtin/builtin_pwd.c
@@ -6,10 +6,11 @@
 /*   By: dsemenov <dsemenov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/05 17:52:23 by dsemenov          #+#    #+#             */
-/*   Updated: 2025/05/07 14:28:58 by dsemenov         ###   ########.fr       */
+/*   Updated: 2025/05/15 16:14:50 by dsemenov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "env.h"
 #include "builtins.h"
 #include <stdio.h>
 #include <stdlib.h>
